Extra tsx refactor
