+++ src/components/navbar/components/Mobile.js
@@ -2,11 +2,11 @@ import { useState } from 'react';
 
 import { Link } from 'react-router-dom';
 
-import BlackArrow from '../../assets/icons/arrow-black.svg';
-import Bars from './img/bars.svg';
-import RainbowLogo from '../../assets/icons/logo-rainbow.png';
+import BlackArrow from '../../../assets/icons/arrow-black.svg';
+import Bars from '../img/bars.svg';
+import RainbowLogo from '../../../assets/icons/logo-rainbow.png';
 
-import styles from './styles/mobile.module.scss';
+import styles from '../styles/mobile.module.scss';
 import classnames from 'classnames';
 
 const SiteNavbar = ({ landing }) => {
